Identity/IdentMetrics split
