+++ libi2pd/TunnelPool.h
@@ -17,6 +17,7 @@
 #include <memory>
 #include <random>
 #include "Identity.h"
+#include "IdentMetrics.h"
 #include "LeaseSet.h"
 #include "RouterInfo.h"
 #include "I2NPProtocol.h"
